HDDS-7294. Move ozone manager background services to a separate package
