+++ hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/ozone/om/multitenant/RangerUserRequest.java
@@ -60,7 +60,7 @@ public class RangerUserRequest {
   private int connectionTimeout = 5000;
   private int connectionRequestTimeout = 5000;
 
-  RangerUserRequest(String rangerHttpsAddress, String userName, String passwd) {
+  public RangerUserRequest(String rangerHttpsAddress, String userName, String passwd) {
 
     // Trim trailing slash
     if (rangerHttpsAddress.endsWith("/")) {
